fix : isbn 포함하는 데이터형으로 수정

+++ flaschenbook-spring-app/flaschenbook/src/main/java/com/book/flaschenbook/entity/BookContentEntity.java
@@ -15,9 +15,7 @@ public class BookContentEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer contentId;
-
-    
+    private Long contentId;
     private String isbn;
     private String content;
 }
